// soulo/src/app/model/profile.ts
import { AbsoluteSourceSpan } from "@angular/compiler";

export class profile{
    userId: number;
    name: string;
    surname: string
    about: string;
    cell: number;
    email: string;
    id: string;
    currentOccupation: string;

    myEducation: string;
    myEducationDescription: string;
    myExperience: string;
    myExperienceDescription : string;
    myProjects: string;
    myProjectsDescription : string; 

    mySkills: string;
    image:string;
    theme:string;

 
  




    
}